+++ webui/src/Controls/Components/LayeredStylesOverrides.tsx
@@ -22,7 +22,7 @@ import { FieldOrExpression } from '~/Components/FieldOrExpression.js'
 import { useComputed } from '~/Resources/util.js'
 import type { IEntityEditorActionService } from '~/Services/Controls/ControlEntitiesService'
 import type { LocalVariablesStore } from '../LocalVariablesStore.js'
-import { OptionsInputControl } from '../OptionsInputField.js'
+import { OptionsInputControl } from '../OptionsInputControl.js'
 import { AddElementPickerModal } from './AddElementPickerModal.js'
 import { ElementPickerModal } from './ElementPickerModal.js'
 import { useLayeredStyleElementsContext } from './LayeredStyleElementsContext.js'
